server/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
Werkzeug==3.0.1
Pillow==10.1.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1